% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auth.R
\name{auth-config}
\alias{auth-config}
\alias{sheets_api_key}
\alias{sheets_oauth_app}
\title{Retrieve OAuth app or API key}
\usage{
sheets_api_key()

sheets_oauth_app()
}
\value{
An \code{\link[httr:oauth_app]{httr::oauth_app()}} or Google API key.
}
\description{
Retrieves the configured OAuth app and API key. Learn more in Google's
document \href{https://support.google.com/googleapi/answer/6158857?hl=en&ref_topic=7013279}{Credentials, access, security, and identity}.
By default, the app and API key are initialized to settings that ship with
googlesheets4. But the user can store their own app or key via
\code{\link[=sheets_auth_config]{sheets_auth_config()}}, i.e. overwrite the defaults.
}
\examples{
sheets_api_key()
sheets_oauth_app()

\dontrun{
sheets_auth_config(api_key = "123")
sheets_api_key()
}
}
\seealso{
Other auth functions: \code{\link{sheets_auth_config}},
  \code{\link{sheets_auth}}, \code{\link{sheets_deauth}}
}
\concept{auth functions}
