% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_drive_user.R
\name{gd_user}
\alias{gd_user}
\alias{gs_user}
\title{Retrieve information about the current Google user}
\usage{
gd_user(verbose = TRUE)

gs_user(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
an object of S3 class \code{drive_user}, which is just a list
}
\description{
Retrieve information about the Google user that has authorized googlesheets
to call the Drive and Sheets APIs on their behalf. Returns info from \href{https://developers.google.com/drive/v2/reference/about/get}{the"about" endpoint}
of the Drive API:
\itemize{
\item User's display name
\item User's email
\item Datetime of the API call
\item User's permission ID
}
}
\examples{
\dontrun{
## these are synonyms: gd = Google Drive, gs = Google Sheets
gd_user()
gs_user()
}

}
\seealso{
Other auth functions: \code{\link{gs_auth}},
  \code{\link{gs_deauth}}
}
