% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{googlesheets4-deprecated}
\alias{googlesheets4-deprecated}
\alias{sheets_sheets}
\alias{sheets_cells}
\alias{sheets_read}
\alias{sheets_write}
\alias{sheets_create}
\alias{sheets_get}
\alias{sheets_example}
\alias{sheets_examples}
\alias{sheets_browse}
\alias{sheets_find}
\alias{sheets_endpoints}
\alias{sheets_auth}
\alias{sheets_deauth}
\alias{sheets_auth_configure}
\alias{sheets_api_key}
\alias{sheets_oauth_app}
\alias{sheets_token}
\alias{sheets_has_token}
\alias{sheets_user}
\title{Deprecated functions}
\usage{
sheets_sheets(...)

sheets_cells(...)

sheets_read(...)

sheets_write(...)

sheets_create(...)

sheets_get(...)

sheets_example(...)

sheets_examples(...)

sheets_browse(...)

sheets_find(...)

sheets_endpoints(...)

sheets_auth(...)

sheets_deauth(...)

sheets_auth_configure(...)

sheets_api_key(...)

sheets_oauth_app(...)

sheets_token(...)

sheets_has_token(...)

sheets_user(...)
}
\arguments{
\item{...}{Passed on to the function that succeeds the deprecated function.}
}
\description{
\lifecycle{deprecated}

These functions are deprecated and will be removed in a future release of
googlesheets4.

\code{sheets_sheets()} is replaced by \code{\link[=sheet_names]{sheet_names()}}.

\code{sheets_cells()} is replaced by \code{\link[=range_read_cells]{range_read_cells()}}.

\code{sheets_read()} is replaced by \code{\link[=range_read]{range_read()}} (which is a synonym for
\code{\link[=read_sheet]{read_sheet()}}).

\code{sheets_write()} is replaced by \code{\link[=sheet_write]{sheet_write()}} (which is a synonym for
\code{\link[=write_sheet]{write_sheet()}}).

\code{sheets_create()} is replaced by \code{\link[=gs4_create]{gs4_create()}}.

\code{sheets_get()} is replaced by \code{\link[=gs4_get]{gs4_get()}}.

\code{sheets_example()} is replaced by \code{\link[=gs4_example]{gs4_example()}}.

\code{sheets_examples()} is replaced by \code{\link[=gs4_examples]{gs4_examples()}}.

\code{sheets_browse()} is replaced by \code{\link[=gs4_browse]{gs4_browse()}}.

\code{sheets_find()} is replaced by \code{\link[=gs4_find]{gs4_find()}}.

\code{sheets_endpoints()} is replaced by \code{\link[=gs4_endpoints]{gs4_endpoints()}}.
}
\section{Auth and API endpoints}{
\tabular{ll}{
   \strong{< v0.2.0} \tab \strong{>= v0.2.0} \cr
   sheets_auth() \tab gs4_auth() \cr
   sheets_auth_configure() \tab gs4_auth_configure() \cr
   sheets_api_key() \tab gs4_api_key() \cr
   sheets_deauth() \tab gs4_deauth() \cr
   sheets_has_token() \tab gs4_has_token() \cr
   sheets_oauth_app() \tab gs4_oauth_app() \cr
   sheets_token() \tab gs4_token() \cr
   sheet_user() \tab gs4_user() \cr
}














}

\keyword{internal}
