% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets4-package.R
\docType{package}
\name{googlesheets4-package}
\alias{googlesheets4}
\alias{googlesheets4-package}
\title{googlesheets4: Access Google Sheets using the Sheets API V4}
\description{
Interact with Google Sheets through the Sheets API
    v4 <https://developers.google.com/sheets/api>. "API" is an acronym for
    "application programming interface"; the Sheets API allows users to
    interact with Google Sheets programmatically, instead of via a web
    browser. The "v4" refers to the fact that the Sheets API is currently
    at version 4. This package helps the user to retrieve Sheet metadata
    and to read data out of specific worksheets or ranges into an R
    object, such as a data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverse/googlesheets4}
  \item Report bugs at \url{https://github.com/tidyverse/googlesheets4/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@rstudio.com} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
