% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets4-package.R, R/gs4_formula.R
\name{googlesheets4-vctrs}
\alias{googlesheets4-vctrs}
\alias{vec_ptype2.googlesheets4_formula}
\alias{vec_cast.googlesheets4_formula}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{googlesheets4_formula}(x, y, ...)

\method{vec_cast}{googlesheets4_formula}(x, to, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
