% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{googlesheets4_formula-vctrs}
\alias{googlesheets4_formula-vctrs}
\alias{vec_ptype2.googlesheets4_formula}
\alias{vec_cast.googlesheets4_formula}
\title{Internal vctrs methods for googlesheets4_formula}
\usage{
\method{vec_ptype2}{googlesheets4_formula}(x, y, ...)

\method{vec_cast}{googlesheets4_formula}(x, to, ...)
}
\description{
Internal vctrs methods for googlesheets4_formula
}
\keyword{internal}
