% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{googlesheets4_formula}
\alias{googlesheets4_formula}
\title{\code{googlesheets4_formula} class}
\usage{
googlesheets4_formula(x = character())
}
\arguments{
\item{x}{Character.}
}
\value{
An S3 vector of class \code{googlesheets4_formula}.
}
\description{
In order to write a formula into Google Sheets, you need to store it as an
object of class \code{googlesheets4_formula}. This is how we distinguish a
"regular" character string from a string that should be interpreted as a
formula. \code{googlesheets4_formula} is an S3 class implemented using the \href{https://vctrs.r-lib.org/articles/s3-vector.html}{vctrs package}.
}
\examples{
if (sheets_has_token()) {
dat <- data.frame(X = c(1, 5, 3, 2, 4, 6))

ss <- sheets_write(dat)

summaries <- tibble::tribble(
       ~ desc, ~ summaries,
        "max", "=max(A:A)",
        "sum", "=sum(A:A)",
        "min", "=min(A:A)",
  "sparkline", "=SPARKLINE(A:A, {\"color\", \"blue\"})"
)

# explicitly declare a column as `googlesheets4_formula`
summaries$summaries <- googlesheets4_formula(summaries$summaries)

sheets_edit(ss, data = summaries, range = "C1", reformat = FALSE)

miscellany <- tibble::tribble(
       ~ desc, ~ example,
  "hyperlink", "=HYPERLINK(\"http://www.google.com/\",\"Google\")",
       "image", "=IMAGE(\"https://www.google.com/images/srpr/logo3w.png\")"
)
miscellany$example <- googlesheets4_formula(miscellany$example)

sheets_write(miscellany, ss = ss)

# clean up
googledrive::drive_trash(ss)
}
}
