% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs4_browse.R
\name{gs4_browse}
\alias{gs4_browse}
\title{Visit a Sheet in a web browser}
\usage{
gs4_browse(ss)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:as_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[=gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[=as_sheets_id]{as_sheets_id()}}.}
}
\value{
The Sheet's browser URL, invisibly.
}
\description{
Visits a Google Sheet in your default browser, if session is interactive.
}
\examples{
gs4_example("mini-gap") \%>\% gs4_browse()
}
