% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs4_example.R
\name{gs4_example}
\alias{gs4_example}
\alias{gs4_examples}
\title{File IDs of example Sheets}
\usage{
gs4_example(matches)

gs4_examples(matches)
}
\arguments{
\item{matches}{A regular expression that matches the nickname of the desired
example Sheet(s). This argument is optional for \code{gs4_examples()} and, if
provided, multiple matches are allowed. \code{gs4_example()} requires
this argument and requires that there is exactly one match.}
}
\value{
\itemize{
\item \code{gs4_example()}: a single \link{sheets_id} object
\item \code{gs4_examples()}: a named vector of all built-in examples, with class
\code{\link[googledrive:as_id]{drive_id}}
}
}
\description{
googlesheets4 ships with static IDs for some world-readable example Sheets
for use in examples and documentation. These functions make them easy to
access by their nicknames.
}
\examples{
gs4_examples()
gs4_examples("gap")
gs4_example("gapminder")
}
