% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs4_example.R
\name{gs4_examples}
\alias{gs4_examples}
\alias{gs4_example}
\title{Example Sheets}
\usage{
gs4_examples(matches)

gs4_example(matches)
}
\arguments{
\item{matches}{A regular expression that matches the name of the desired
example Sheet(s). \code{matches} is optional for the plural \code{gs4_examples()}
and, if provided, it can match multiple Sheets. The singular
\code{gs4_example()} requires \code{matches} and it must match exactly one Sheet.}
}
\value{
\itemize{
\item \code{gs4_example()}: a \link{sheets_id}
\item \code{gs4_examples()}: a named vector of all built-in examples, with class
\code{\link[googledrive:as_id]{drive_id}}
}
}
\description{
googlesheets4 makes a variety of world-readable example Sheets available for
use in documentation and reprexes. These functions help you access the
example Sheets. See \code{vignette("example-sheets", package = "googlesheets4")}
for more.
}
\examples{
\dontshow{if (gs4_has_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gs4_examples()
gs4_examples("gap")

gs4_example("gapminder")
gs4_example("deaths")
\dontshow{\}) # examplesIf}
}
