% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs4_auth.R
\name{gs4_has_token}
\alias{gs4_has_token}
\title{Is there a token on hand?}
\usage{
gs4_has_token()
}
\value{
Logical.
}
\description{
Reports whether googlesheets4 has stored a token, ready for use in downstream
requests.
}
\examples{
gs4_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{gs4_token}()},
\code{\link{request_generate}()},
\code{\link{request_make}()}
}
\concept{low-level API functions}
