% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_make_request.R
\name{gs_make_request}
\alias{gs_make_request}
\title{Execute a request}
\usage{
gs_make_request(req)
}
\arguments{
\item{req}{a request as prepared by \code{\link[=gs_generate_request]{gs_generate_request()}} or
\code{\link[=gs_build_request]{gs_build_request()}}}
}
\value{
the response
}
\description{
Placeholder so links from elsewhere aren't broken.
}
