% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dimension.R
\name{letter_to_num2}
\alias{letter_to_num2}
\title{Convert character column specifications to groups of numeric indices.}
\usage{
letter_to_num2(indexes)
}
\arguments{
\item{indexes}{The character column specification ex. C:F}
}
\value{
\code{(list)} A list of numeric vectors with the column numbers. If a series of column numbers the starting and ending index are returned.
}
\description{
This differs from cellranger::letter_to_num in that it accounts for series of columns ie B:D and the output is a list.
}
\examples{
\donttest{
indexes <- c("a")
letter_to_num2(indexes)
indexes <- c("c:f", "R:S", "AB:AC")
letter_to_num2(indexes)
}
}
