% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet-print.R
\name{print.gsheet}
\alias{print.gsheet}
\title{Print info about a \code{gsheet} object}
\usage{
\method{print}{gsheet}(x, ...)
}
\arguments{
\item{x}{\code{\link{gsheet}} object returned by functions such as \code{\link[=sheet_name]{sheet_name()}},
\code{\link[=sheet_id]{sheet_id()}}, and friends}

\item{...}{potential further arguments (required for Method/Generic reasons)}
}
\description{
Display information about a Google spreadsheet that has been registered with
googlesheets: the name or title of the spreadsheet, date-time of
registration, date-time of last update (at time of registration), visibility,
permissions, version, the number of worksheets contained, worksheet titles
and extent, and sheet key. \emph{A lot of that is not true yet or anymore!}
}
\examples{
\dontrun{
foo <- gs_new("foo")
foo
print(foo)
}

}
