% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_format.R
\name{range_format_pattern}
\alias{range_format_pattern}
\title{Range format following pattern}
\usage{
range_format_pattern(ss, pattern, sheet = NULL, range = NULL)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=gs4_get]{gs4_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{pattern}{Character. Google Sheets format pattern. For options see the
\href{https://developers.google.com/sheets/api/guides/formats}{Google Documentation}.}

\item{sheet}{Sheet to modify, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via \code{range}. If neither argument specifies the sheet, defaults to the first visible sheet.}

\item{range}{Which columns or rows to resize. Optional.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Applies Google Sheets number formatting to control how your data appears in a
sheet for a given range.
}
\examples{
if (gs4_has_token()) {
  dat <- data.frame(small_number = runif(12),
                    big_number = runif(12) * 1e6,
                    date = Sys.Date() - (0:11) * 30)
  # make the sheet
  ss <- gs4_create("gs4-number-formats-demo", sheets = dat)
  # format as percent
  range_format_pattern(ss, range = "A", pattern = "0.0\%")
  # format as big number
  # from https://webapps.stackexchange.com/questions/77974
  range_format_pattern(ss, range = "B", pattern = "[>999999]0.0,,\\\\M;[>999]0.0,\\\\K;0")
  # format as day-month
  range_format_pattern(ss, range = "C", pattern = "dd\"-\"mmmm")

  # clean up
  gs4_find("gs4-number-formats-demo") \%>\%
    googledrive::drive_trash()
}
}
\seealso{
Other formatting functions: 
\code{\link{range_autofit}()}
}
\concept{formatting functions}
