% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheet.R
\name{read_sheet}
\alias{read_sheet}
\title{Read a Sheet into a data frame}
\usage{
read_sheet(ss, sheet = NULL, range = NULL, col_names = TRUE,
  col_types = NULL, na = "", trim_ws = TRUE, skip = 0, n_max = Inf,
  guess_max = min(1000, n_max))
}
\arguments{
\item{ss}{Something that uniquely identifies a Google Sheet. Processed
through \code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
visible sheet. \emph{wording basically copied from readxl}}

\item{range}{A cell range to read from, as described in FILL THIS IN
\emph{wording basically copied copied from readxl}}

\item{col_names}{column names}

\item{col_types}{column types}

\item{na}{na strings}

\item{trim_ws}{whether to trim ws}

\item{skip}{rows to skip}

\item{n_max}{max data rows to read}

\item{guess_max}{max rows to consult in column typing}
}
\value{
a tibble
}
\description{
WIP! The main read function of this package.
}
\examples{
read_sheet(sheets_example("mini-gap"))
test_sheet <- "1J5gb0u8n3D2qx3O3rY28isnI5SD89attRwhWPWlkmDM"
read_sheet(test_sheet)
read_sheet(test_sheet, skip = 2)
read_sheet(test_sheet, n_max = 2)
read_sheet(test_sheet, range = "A1:B2")
read_sheet(test_sheet, range = "B2:C4")
read_sheet(test_sheet, range = "B2:E5")

ss <- sheets_example("deaths")
range <- "A5:F15"
col_types <- "ccilDD"
read_excel(readxl_example("deaths.xlsx"), range = "other!A5:F15")
read_sheet(ss, range = "other!A5:F15", col_types = "ccilDD")
}
