% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_process.R
\name{response_process}
\alias{response_process}
\title{Process a response from a Google API}
\usage{
response_process(resp)
}
\arguments{
\item{resp}{Object of class \code{response} from \link{httr}.}
}
\value{
List.
}
\description{
Temporarily gestating this in googlesheets4 but it is destined for gargle.
}
\seealso{
Other low-level API functions: \code{\link{request_generate}},
  \code{\link{request_make}}
}
\concept{low-level API functions}
