% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_append.R
\name{sheets_append}
\alias{sheets_append}
\title{Append rows to a sheet}
\usage{
sheets_append(data, ss, sheet = 1)
}
\arguments{
\item{data}{A data frame.}

\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to append to, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Adds one or more new rows after the last row with data in a (work)sheet,
increasing the row dimension of the sheet if necessary.
}
\examples{
if (sheets_has_token()) {
  # we will recreate the table of "other" deaths from this example Sheet
  (deaths <- sheets_example("deaths") \%>\%
    sheets_read(range = "other_data", col_types = "????DD"))

  # split the data into 3 pieces, which we will send separately
  deaths_one <- deaths[1:5, ]
  deaths_two <- deaths[6, ]
  deaths_three <- deaths[7:10, ]

  # create a Sheet and send the first chunk of data
  ss <- sheets_create("sheets-append-demo", sheets = list(deaths = deaths_one))

  # append a single row
  sheets_append(deaths_two, ss)

  # append remaining rows
  sheets_append(deaths_three, ss)

  # read and check against the original
  deaths_replica <- sheets_read(ss, col_types = "????DD")
  identical(deaths, deaths_replica)

  # cleanup
  googledrive::drive_rm(ss)
}
}
\seealso{
Makes an \code{AppendCellsRequest}:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#AppendCellsRequest}
}

Other write functions: 
\code{\link{sheets_edit}()},
\code{\link{write_sheet}()}
}
\concept{write functions}
