% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auth.R
\name{sheets_auth}
\alias{sheets_auth}
\title{Authorize googlesheets4}
\usage{
sheets_auth(email = NULL, path = NULL,
  scopes = "https://www.googleapis.com/auth/spreadsheets",
  cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default())
}
\arguments{
\item{email}{Optional; email address associated with the desired Google user.}

\item{path}{Optional; path to the downloaded JSON file for a service token.}

\item{scopes}{Optional; scope(s) to use. See your choices at
\href{https://developers.google.com/identity/protocols/googlescopes#sheetsv4}{OAuth 2.0 Scopes for Google APIs}.}

\item{cache}{Specifies the OAuth token cache. Defaults to the option named
"gargle_oauth_cache", retrieved via \code{\link[=gargle_oauth_cache]{gargle_oauth_cache()}}.}

\item{use_oob}{Whether to prefer "out of band" authentication. Defaults to
the option named "gargle_oob_default", retrieved via
\code{\link[=gargle_oob_default]{gargle_oob_default()}}.}
}
\description{
Authorize googlesheets4 to view and manage your Google Sheets. By default,
you are directed to a web browser, asked to sign in to your Google
account, and to grant googlesheets4 permission to operate on your behalf
with Google Sheets. By default, these user credentials are cached in a
folder below your home directory, \code{~/.R/gargle/gargle-oauth}, from where
they can be automatically refreshed, as necessary. Storage at the user
level means the same token can be used across multiple projects and
tokens are less likely to be synced to the cloud by accident.
}
\details{
Most users, most of the time, do not need to call \code{sheets_auth()}
explicitly -- it is triggered by the first action that requires
authorization. Even when called, the default arguments often suffice.
However, when necessary, this function allows the user to explicitly:
\itemize{
\item Declare which Google identity to use, via an email address. If there
are multiple cached tokens, this can clarify which one to use. It can
also force googlesheets4 to switch from one identity to another. If
there's no cached token for the email, this triggers a return to the
browser to choose the identity and give consent.
\item Use a service account token.
\item Specify non-default behavior re: token caching and out-of-bound
authentication.
}

For even deeper control over auth, use \code{\link[=sheets_auth_config]{sheets_auth_config()}} to use
your own OAuth app or API key.
}
\examples{
\dontrun{
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
sheets_auth()

## force use of a token associated with a specific email
sheets_auth(email = "jenny@example.com")

## use a 'read only' scope, so it's impossible to edit or delete Sheets
sheets_auth(
  scopes = "https://www.googleapis.com/auth/spreadsheets.readonly"
)

## use a service account token
sheets_auth(path = "foofy-83ee9e7c9c48.json")
}
}
\seealso{
Other auth functions: \code{\link{sheets_auth_config}},
  \code{\link{sheets_deauth}}
}
\concept{auth functions}
