% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auto_resize_dims.R
\name{sheets_auto_resize_dims}
\alias{sheets_auto_resize_dims}
\title{Auto-resize columns or rows}
\usage{
sheets_auto_resize_dims(
  ss,
  sheet = NULL,
  range = NULL,
  dimension = c("columns", "rows")
)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to modify, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via \code{range}. If neither argument specifies the sheet, defaults to the first visible sheet.}

\item{range}{Which columns or rows to resize. Optional. If you want to resize
all columns or all rows, use \code{dimension} instead. All the usual \code{range}
specifications are accepted, but the targeted range must specify only
columns (e.g. "B:F") or only rows (e.g. "2:7").}

\item{dimension}{Ignored if \code{range} is given. If consulted, \code{dimension} must
be either \code{"columns"} (the default) or \code{"rows"}. This is the simplest way
to request auto-resize for all columns or all rows.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Applies automatic resizing to either columns or rows of a (work)sheet. The
width or height of targeted columns or rows, respectively, is determined
from the current cell contents. This only affects the appearance of a Sheet
in the browser and doesn't affect its values in any way.
}
\examples{
if (sheets_has_token()) {
dat <- tibble::tibble(
  fruit = c("date", "lime", "pear", "plum")
)

ss <- sheets_write(dat)
ss

# open in the browser
sheets_browse(ss)

# shrink column A to fit the short fruit names
sheets_auto_resize_dims(ss)

# send some longer fruit names
dat2 <- tibble::tibble(
  fruit = c("cucumber", "honeydew")
)
ss \%>\% sheets_append(dat2)
# in the browser, see that column A is now too narrow to show the data

sheets_auto_resize_dims(ss)
# in the browser, see the column A reveals all the data now

googledrive::drive_trash(ss)
}
}
\seealso{
Makes an \code{AutoResizeDimensionsRequest}:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#autoresizedimensionsrequest}
}
}
\concept{formatting functions}
