% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_browse.R
\name{sheets_browse}
\alias{sheets_browse}
\title{Visit Sheet in browser}
\usage{
sheets_browse(ss)
}
\arguments{
\item{ss}{Something that uniquely identifies a Google Sheet. Processed
through \code{\link[=as_sheets_id]{as_sheets_id()}}.}
}
\value{
Character vector of file hyperlinks, from
\code{\link[googledrive:drive_link]{googledrive::drive_link()}}, invisibly.
}
\description{
Visits a Google Sheet in your default browser. TODO: Note that there is no
provision for auth yet, so will only work for a Sheet that is readable to
anyone.
}
\examples{
\dontrun{
sheets_example("mini-gap") \%>\% sheets_browse()
}
}
