% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_create.R
\name{sheets_create}
\alias{sheets_create}
\title{Create a new Sheet}
\usage{
sheets_create(name, ..., sheets = NULL)
}
\arguments{
\item{name}{The name of the new spreadsheet.}

\item{...}{Optional spreadsheet properties that can be set through this API
endpoint, such as locale and time zone.}

\item{sheets}{Optional named list of data frames. One sheet is created for
each data frame.}
}
\value{
A list with S3 class \code{googlesheets4_spreadsheet}, for printing
purposes.
}
\description{
\lifecycle{experimental}

Creates an entirely new (spread)Sheet (or, in Excel-speak, workbook). Offers
some control over the initial set of (work)sheets (a.k.a. "tabs").
}
\examples{
if (sheets_has_token()) {
  sheets_create("sheets-create-demo-1")

  sheets_create("sheets-create-demo-2", locale = "en_CA")

  sheets_create(
    "sheets-create-demo-3",
    locale = "fr_FR",
    timeZone = "Europe/Paris"
  )

  sheets_create(
    "sheets-create-demo-4",
    sheets = list(iris = head(iris), mtcars = head(mtcars))
  )

  # clean up
  sheets_find("sheets-create-demo") \%>\% googledrive::drive_trash()
}
}
\seealso{
Wraps the \code{spreadsheets.create} endpoint:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/create}
}
}
