% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auth.R
\name{sheets_deauth}
\alias{sheets_deauth}
\title{Suspend authorization}
\usage{
sheets_deauth()
}
\description{
Put googlesheets4 into a de-authorized state. Instead of sending a token,
googlesheets4 will send an API key. This can be used to access public
resources for which no Google sign-in is required. This is handy for using
googlesheets4 in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. A built-in API key is used by default or
the user can configure their own via \code{\link[=sheets_auth_config]{sheets_auth_config()}}.
}
\examples{
\dontrun{
sheets_deauth()
sheets_email()

# get metadata on the public 'deaths' spreadsheet
sheets_get("1ESTf_tH08qzWwFYRC1NVWJjswtLdZn9EGw5e3Z5wMzA")
}
}
\seealso{
Other auth functions: \code{\link{sheets_auth_config}},
  \code{\link{sheets_auth}}
}
\concept{auth functions}
