% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_email.R
\name{sheets_email}
\alias{sheets_email}
\title{Get email of current user}
\usage{
sheets_email()
}
\value{
An email address or \code{NULL}.
}
\description{
Reveals email associated with the Google credential currently being used by
googlesheets4 (if there is one). This is semi experimental, inspired by
\code{\link[googledrive:drive_user]{googledrive::drive_user()}}. The fate of this function to be determined by
how scopes shake out and whether we create a way for googledrive and
googlesheets4 to share a token.
}
\examples{
sheets_email()
}
