% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auth.R
\name{sheets_has_token}
\alias{sheets_has_token}
\title{Is there a token on hand?}
\usage{
sheets_has_token()
}
\value{
Logical.
}
\description{
Reports whether googlesheets4 has stored a token, ready for use in downstream
requests. Exists mostly for protecting examples that won't work in the
absence of a token.
}
\examples{
sheets_has_token()
}
