% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_id.R
\name{sheets_id}
\alias{sheets_id}
\title{sheets_id object}
\description{
Holds a spreadsheet identifier. This is what the Sheets and Drive APIs refer
to as \code{spreadsheetId} and \code{fileId}, respectively.
}
\seealso{
\code{\link[=as_sheets_id]{as_sheets_id()}}
}
