% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_list.R
\name{sheets_list}
\alias{sheets_list}
\title{List your Google Sheets}
\usage{
sheets_list(...)
}
\arguments{
\item{...}{Arguments (other than \code{type}, which is hard-wired as \code{type = "spreadsheet"}) that are passed along to \code{\link[googledrive:drive_find]{googledrive::drive_find()}}.}
}
\value{
An object of class \code{\link[googledrive]{dribble}}, a tibble with one row per item.
}
\description{
Lists your Google Sheets. This is a very thin wrapper around
\code{\link[googledrive:drive_find]{googledrive::drive_find()}}, that specifies you want to list Drive files
where \code{type = "spreadsheet"}. Therefore, note that this will require auth
for googledrive!
}
\examples{
if (interactive()) {
  # see all your Sheets
  sheets_list()

  # see 5 Sheets, prioritized by creation time
  x <- sheets_list(order_by = "createdTime desc", n_max = 5)
  x

  # hoist the creation date, using other packages in the tidyverse
  # x \%>\%
  #   tidyr::hoist(drive_resource, created_on = "createdTime") \%>\%
  #   dplyr::mutate(created_on = as.Date(created_on))
}
}
