% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_create.R
\name{sheets_random}
\alias{sheets_random}
\title{Generate a random Sheet name}
\usage{
sheets_random(n = 1)
}
\arguments{
\item{n}{Number of names to generate.}
}
\value{
A character vector.
}
\description{
Generates a random name, suitable for a newly created Sheet, using
\code{\link[ids:adjective_animal]{ids::adjective_animal()}}.
}
\examples{
sheets_random()
}
