% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_add.R
\name{sheets_sheet_add}
\alias{sheets_sheet_add}
\title{Add a (work)sheet}
\usage{
sheets_sheet_add(ss, sheet = NULL, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{A string providing the new sheet's name. If unspecified, Sheets
autogenerates a name of the form "SheetN".}

\item{...}{Optional parameters to specify additional properties of the new
sheet. Not relevant to most users. Specify fields of the \href{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/sheets#SheetProperties}{\code{SheetProperties} schema}
in \code{name = value} form.}

\item{.before, .after}{Optional specification of where to put the new sheet.
Can be an existing sheet name or a position. If unspecified, Sheets puts
the new sheet at the end.}
}
\value{
\code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Adds a (work)sheet to an existing (spread)Sheet.
}
\examples{
if (sheets_has_token()) {
ss <- sheets_create("add-sheets-to-me")

# the only required argument is the target spreadsheet
ss \%>\% sheets_sheet_add()

# but you CAN specify sheet name and/or position
ss \%>\% sheets_sheet_add("apple", .after = 1)
ss \%>\% sheets_sheet_add("banana", .after = "apple")

# keeners can even specify additional sheet properties
ss \%>\%
sheets_sheet_add(
  sheet = "coconut",
  gridProperties = list(
              rowCount = 3, columnCount = 6, frozenRowCount = 1
              )
              )

# get an overview of the sheets
sheets_sheet_data(ss)

# cleanup
sheets_find("add-sheets-to-me") \%>\% googledrive::drive_rm()
}
}
\seealso{
Makes an \code{AddSheetsRequest}:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#addsheetrequest}
}
}
