% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_copy.R
\name{sheets_sheet_copy}
\alias{sheets_sheet_copy}
\title{Copy a (work)sheet}
\usage{
sheets_sheet_copy(
  from_ss,
  from_sheet = NULL,
  to_ss = from_ss,
  to_sheet = NULL,
  .before = NULL,
  .after = NULL
)
}
\arguments{
\item{from_ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{from_sheet}{Sheet to copy, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Defaults to the first visible sheet.}

\item{to_ss}{The Sheet to copy \emph{to}. Accepts all the same types of input as
\code{from_ss}, which is also what this defaults to, if unspecified.}

\item{to_sheet}{Not implemented yet. But once it is: Name of the new sheet,
as a string. If you don't specify this, Google generates a name, along the
line of "Copy of blah". Note that sheet names must be unique within a
Sheet, so if the automatic name would violate this, Google also
de-duplicates it for you, e.g. "Copy of blah 2". If you have strong
opinions about these matters, you should specify \code{to_sheet}.}

\item{.before}{Not implemented yet.}

\item{.after}{Not implemented yet.}
}
\value{
The receiving Sheet, \verb{to_ ss}, as an instance of \code{\link{sheets_id}}.
}
\description{
Copies a (work)sheet, within its current (spread)Sheet or to another Sheet.
}
\examples{
if (sheets_has_token()) {
  ss_aaa <- sheets_create(
    "sheets-sheet-copy-demo-aaa",
    sheets = list(iris = head(iris), chickwts = head(chickwts))
  )

  # copies 'iris' sheet within existing Sheet
  ss_aaa \%>\%
    sheets_sheet_copy()

  # make a second Sheet
  ss_bbb <- sheets_create("sheets-sheet-copy-demo-bbb")

  # copies 'chickwts' sheet from first Sheet to second
  ss_aaa \%>\%
    sheets_sheet_copy("chickwts", to_ss = ss_bbb)

  # clean up
  googledrive::drive_find("sheets-sheet-copy-demo") \%>\%
    googledrive::drive_trash()
}
}
\seealso{
Makes a \code{copyTo} request:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.sheets/copyTo}
}

Other worksheet functions: 
\code{\link{sheets_sheet_add}()},
\code{\link{sheets_sheet_delete}()},
\code{\link{sheets_sheet_properties}()},
\code{\link{sheets_sheet_resize}()}
}
\concept{worksheet functions}
