% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_delete.R
\name{sheets_sheet_delete}
\alias{sheets_sheet_delete}
\title{Delete one or more (work)sheets}
\usage{
sheets_sheet_delete(ss, sheet)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to delete, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. You can pass a vector to delete multiple sheets at once or even a list, if you need to mix names and positions.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Deletes one or more (work)sheets from a (spread)Sheet.
}
\examples{
if (sheets_has_token()) {
  ss <- sheets_create("delete-sheets-from-me")
  sheets_sheet_add(ss, c("alpha", "beta", "gamma", "delta"))

  # get an overview of the sheets
  sheets_sheet_properties(ss)

  # delete sheets
  sheets_sheet_delete(ss, 1)
  sheets_sheet_delete(ss, "gamma")
  sheets_sheet_delete(ss, list("alpha", 2))

  # get an overview of the sheets
  sheets_sheet_properties(ss)

  # cleanup
  sheets_find("delete-sheets-from-me") \%>\% googledrive::drive_rm()
}
}
\seealso{
Makes an \code{DeleteSheetsRequest}:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#DeleteSheetRequest}
}

Other worksheet functions: 
\code{\link{sheet_append}()},
\code{\link{sheets_sheet_add}()},
\code{\link{sheets_sheet_copy}()},
\code{\link{sheets_sheet_properties}()},
\code{\link{sheets_sheet_relocate}()},
\code{\link{sheets_sheet_rename}()},
\code{\link{sheets_sheet_resize}()}
}
\concept{worksheet functions}
