% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_properties.R
\name{sheets_sheet_properties}
\alias{sheets_sheet_properties}
\alias{sheets_sheet_names}
\title{Get data about (work)sheets}
\usage{
sheets_sheet_properties(ss)

sheets_sheet_names(ss)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}
}
\value{
\itemize{
\item \code{sheets_sheet_properties()}: A tibble with one row per (work)sheet.
\item \code{sheets_sheet_names()}: A character vector of (work)sheet names.
}
}
\description{
Reveals full metadata or just the names for the (work)sheets inside a
(spread)Sheet.
}
\examples{
if (sheets_has_token()) {
  ss <- sheets_example("gapminder")
  sheets_sheet_properties(ss)
  sheets_sheet_names(ss)
}
}
\seealso{
Other worksheet functions: 
\code{\link{sheets_sheet_add}()},
\code{\link{sheets_sheet_copy}()},
\code{\link{sheets_sheet_delete}()},
\code{\link{sheets_sheet_relocate}()},
\code{\link{sheets_sheet_rename}()},
\code{\link{sheets_sheet_resize}()}
}
\concept{worksheet functions}
