% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_relocate.R
\name{sheets_sheet_relocate}
\alias{sheets_sheet_relocate}
\title{Relocate one or more (work)sheets}
\usage{
sheets_sheet_relocate(ss, sheet, .before = NULL, .after = NULL)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to relocate, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. You can pass a vector to move multiple sheets at once or even a list, if you need to mix names and positions.}

\item{.before, .after}{Optional specification of where to put the new sheet. Specify, at most, one of \code{.before} and \code{.after}. Refer to an existing sheet by name (via a string) or by position (via a number). If unspecified, Sheets puts the new sheet at the end.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Move (work)sheets around within a (spread)Sheet. The results are most
predictable for these common and simple use cases:
\itemize{
\item Move a single sheet.
\item Move multiple sheets to the front with \code{.before = 1}.
\item Move multiple sheets to the back with \code{.after = 100} (\code{.after} can be
any number greater than or equal to the number of sheets).
\item Exhaustively list existing sheets in a new order and specify .\code{before = 1}.
}

If your relocating is more complicated and you are puzzled by the results,
express it as a sequence of simpler calls to \code{sheets_sheet_relocate()}.
}
\examples{
if (sheets_has_token()) {
  ss <- sheets_create(
    "sheets-sheet-relocate-demo",
    sheets = c("alfa", "bravo", "charlie", "delta", "echo", "foxtrot")
  )
  sheets_sheet_names(ss)

  # move one sheet, forwards then backwards
  ss \%>\%
    sheets_sheet_relocate("echo", .before = "bravo") \%>\%
    sheets_sheet_names()
  ss \%>\%
    sheets_sheet_relocate("echo", .after = "delta") \%>\%
    sheets_sheet_names()

  # move multiple sheets to the front
  ss \%>\%
    sheets_sheet_relocate(list("foxtrot", 4), .before = 1) \%>\%
    sheets_sheet_names()

  # exhaustively list all sheets in a desired order
  new_order <- rev(c("alfa", "bravo", "charlie", "delta", "echo", "foxtrot"))
  ss \%>\%
    sheets_sheet_relocate(new_order, .before = 1) \%>\%
    sheets_sheet_names()

  # move multiple sheets to the back
  ss \%>\%
    sheets_sheet_relocate(c("bravo", "alfa", "echo"), .after = 10) \%>\%
    sheets_sheet_names()

  # clean up
  googledrive::drive_find("sheets-sheet-relocate-demo") \%>\%
    googledrive::drive_trash()
}
}
\seealso{
Constructs a batch of \code{UpdateSheetPropertiesRequest}s (one per sheet):
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#UpdateSheetPropertiesRequest}
}

Other worksheet functions: 
\code{\link{sheets_sheet_add}()},
\code{\link{sheets_sheet_copy}()},
\code{\link{sheets_sheet_delete}()},
\code{\link{sheets_sheet_properties}()},
\code{\link{sheets_sheet_rename}()},
\code{\link{sheets_sheet_resize}()}
}
\concept{worksheet functions}
