% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_rename.R
\name{sheets_sheet_rename}
\alias{sheets_sheet_rename}
\title{Rename one  (work)sheet}
\usage{
sheets_sheet_rename(ss, sheet, new_sheet)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to rename, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. You can pass a sheet name or position.}

\item{new_sheet}{New name to rename the sheet to.}
}
\value{
A list containing the name and id of the renamed sheet.
}
\description{
Rename one (work)sheet from a (spread)Sheet
}
\examples{
if (sheets_has_token()) {
  ss <- sheets_create("rename-sheets-from-me")
  sheets_sheet_add(ss, c("alpha", "beta", "gamma"))

  # get an overview of the sheets
  sheets_sheet_data(ss)

  # rename sheets
  sheets_sheet_rename(ss, 1,"alpha particle")
  sheets_sheet_rename(ss, "gamma","gamma ray")

  # get an overview of the sheets
  sheets_sheet_data(ss)

  # cleanup
  sheets_find("rename-sheets-from-me") \%>\% googledrive::drive_rm()
}
}
\seealso{
Makes an \code{UpdateSheetPropertiesRequest}:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#UpdateSheetPropertiesRequest}
}

Other worksheet functions: 
\code{\link{sheets_sheet_add}()},
\code{\link{sheets_sheet_copy}()},
\code{\link{sheets_sheet_data}()},
\code{\link{sheets_sheet_delete}()}
}
\concept{worksheet functions}
