% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_resize.R
\name{sheets_sheet_resize}
\alias{sheets_sheet_resize}
\title{Change the size of a (work)sheet}
\usage{
sheets_sheet_resize(ss, sheet = NULL, nrow = NULL, ncol = NULL, exact = FALSE)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to resize, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number.}

\item{nrow, ncol}{Desired number of rows or columns, respectively. The default
of \code{NULL} means to leave unchanged.}

\item{exact}{Logical, indicating whether to impose \code{nrow} and \code{ncol} exactly
or to treat them as lower bounds. If \code{exact = FALSE},
\code{sheets_sheet_resize()} can only add cells. If \code{exact = TRUE}, cells can be
deleted and their contents are lost.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Changes the number of rows and/or columns in a (work)sheet.
}
\examples{
if (sheets_has_token()) {
  # create a Sheet with the default initial worksheet
  (ss <- sheets_create("sheets-resize-demo"))

  # see (work)sheet dims
  sheets_sheet_properties(ss)

  # no resize occurs
  sheets_sheet_resize(ss, nrow = 2, ncol = 6)

  # reduce sheet size
  sheets_sheet_resize(ss, nrow = 5, ncol = 7, exact = TRUE)

  # add rows
  sheets_sheet_resize(ss, nrow = 7)

  # add columns
  sheets_sheet_resize(ss, ncol = 10)

  # add rows and columns
  sheets_sheet_resize(ss, nrow = 9, ncol = 12)

  # re-inspect (work)sheet dims
  sheets_sheet_properties(ss)

  googledrive::drive_rm(ss)
}
}
\seealso{
Makes an \code{UpdateSheetPropertiesRequest}:
\itemize{
\item <# https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/request#UpdateSheetPropertiesRequest>
}

Other worksheet functions: 
\code{\link{sheet_append}()},
\code{\link{sheets_sheet_add}()},
\code{\link{sheets_sheet_copy}()},
\code{\link{sheets_sheet_delete}()},
\code{\link{sheets_sheet_properties}()},
\code{\link{sheets_sheet_relocate}()},
\code{\link{sheets_sheet_rename}()}
}
\concept{worksheet functions}
