% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_update_cells.R
\name{sheets_sheet_update_cells}
\alias{sheets_sheet_update_cells}
\title{Update cell(s) content}
\usage{
sheets_sheet_update_cells(
  df,
  ss,
  sheet,
  valueInputOption = c("RAW", "USER_ENTERD")
)
}
\arguments{
\item{df}{dataframe with columns \code{loc} for A1 type location and \code{value} with value to update with}

\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to update}

\item{valueInputOption}{optional, determines if input is raw or should be parsed}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
Update the content of individual cell(s)
}
