% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auth.R
\name{sheets_user}
\alias{sheets_user}
\title{Get info on current user}
\usage{
sheets_user()
}
\value{
An email address or, if no token has been loaded, \code{NULL}.
}
\description{
Reveals the email address of the user associated with the current token.
If no token has been loaded yet, this function does not initiate auth.
}
\examples{
\dontrun{
sheets_user()
}
}
\seealso{
\code{\link[gargle:token_userinfo]{gargle::token_userinfo()}}, \code{\link[gargle:token_email]{gargle::token_email()}},
\code{\link[gargle:token_tokeninfo]{gargle::token_tokeninfo()}}
}
