% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dimension.R
\name{split_consecutive}
\alias{split_consecutive}
\title{Convert non-consecutive indexes to consecutive groupings}
\usage{
split_consecutive(indexes)
}
\arguments{
\item{indexes}{\code{(numeric)} The numeric vector to group}
}
\value{
\code{(list)} A list of consecutive groupings of integer indexes
}
\description{
Convert non-consecutive indexes to consecutive groupings
}
\examples{
\donttest{
split_consecutive(c(1:2, 5:6))
}
}
