% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_write.R
\name{write_sheet}
\alias{write_sheet}
\alias{sheets_write}
\title{(Over)write new data into an existing sheet}
\usage{
write_sheet(data, ss, sheet = NULL, skip = 0, na = "")

sheets_write(data, ss, sheet = NULL, skip = 0, na = "")
}
\arguments{
\item{data}{A data frame.}

\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to write into, as in "worksheet" or "tab". Either a string
(the name of a sheet), or an integer (the position of the sheet). If
unspecified, defaults to the first sheet.}

\item{skip}{Number of rows to leave empty before starting to write.}

\item{na}{Not implemented yet.}
}
\value{
Updated metadata for the (spread)Sheet \code{ss}, as an instance of S3
class \code{googlesheets4_spreadsheet}.
}
\description{
Updates an existing (work)sheet in an existing (spread)Sheet with a data
frame. All pre-existing values, formats, and dimensions of the targeted sheet
are cleared and it gets its new values and dimensions from \code{data}. Special
formatting is applied to the header row, which holds column names, and the
first \code{skip + 1} rows are frozen (so, up to and including the header row).
\emph{NOTE: this function is very alpha and currently writes everything as
character.}
}
\examples{
if (sheets_has_token()) {
  # create a Sheet with some initial, placeholder data
  ss <- sheets_create(
    "sheets-write-demo",
    sheets = list(alpha = data.frame(x = 1), omega = data.frame(x = 1))
  )

  df <- data.frame(
    x = 1:3,
    y = letters[1:3],
    stringsAsFactors = FALSE
  )

  # write df into the first sheet
  (ss <- sheets_write(data = df, ss = ss))

  # write mtcars into the sheet named 'omega'
  (ss <- sheets_write(data = mtcars, ss = ss, sheet = "omega"))

  # view your magnificent creation in the browser
  # sheets_browse(ss)

  # clean up
  sheets_find("sheets-write-demo") \%>\% googledrive::drive_rm()
}
}
