% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dimension.R
\name{delete_dimension}
\alias{delete_dimension}
\title{Delete rows or columns of a sheet.}
\usage{
delete_dimension(
  ss = NULL,
  sheet = NULL,
  dimension = c("ROWS", "COLUMNS"),
  indexes = NULL
)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{\code{(character/numeric)} The name of the sheet (case-sensitive) or the index (where the first sheet is 1.)}

\item{dimension}{\code{(character)} The dimension to delete ROWS or COLUMNS. Can be shorthand: ex. 'r' or 'c'}

\item{indexes}{\code{(character/numeric)} The row numbers, column numbers or column names (in C:F format) to delete.}
}
\value{
Returns the supplied spreadsheet object invisibly
}
\description{
\lifecycle{experimental}

Delete rows or columns of a specified sheet given a vector of row numbers, column names, or column numbers.
}
\examples{
\dontrun{
delete_dimension(ss, "Sheet1", "c", c("A", "D:E"))
delete_dimension(ss, "Sheet1", "r", c(1, 3:5))
}
}
