% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{googlesheets4-deprecated}
\alias{googlesheets4-deprecated}
\alias{sheets_sheets}
\alias{sheets_cells}
\alias{sheets_read}
\alias{sheets_write}
\alias{sheets_create}
\alias{sheets_get}
\alias{sheets_example}
\alias{sheets_examples}
\alias{sheets_browse}
\alias{sheets_find}
\alias{sheets_endpoints}
\title{Deprecated functions}
\usage{
sheets_sheets(...)

sheets_cells(...)

sheets_read(...)

sheets_write(...)

sheets_create(...)

sheets_get(...)

sheets_example(...)

sheets_examples(...)

sheets_browse(...)

sheets_find(...)

sheets_endpoints(...)
}
\arguments{
\item{...}{Passed on to the function that succeeds the deprecated function.}
}
\description{
\lifecycle{deprecated}

These functions are deprecated and will be removed in a future release of
googlesheets4.

\code{sheets_sheets()} is replaced by \code{\link[=sheet_names]{sheet_names()}}.

\code{sheets_cells()} is replaced by \code{\link[=range_read_cells]{range_read_cells()}}.

\code{sheets_read()} is replaced by \code{\link[=range_read]{range_read()}} (which is a synonym for
\code{\link[=read_sheet]{read_sheet()}}).

\code{sheets_write()} is replaced by \code{\link[=sheet_write]{sheet_write()}} (which is a synonym for
\code{\link[=write_sheet]{write_sheet()}}).

\code{sheets_create()} is replaced by \code{\link[=gs4_create]{gs4_create()}}.

\code{sheets_get()} is replaced by \code{\link[=gs4_get]{gs4_get()}}.

\code{sheets_example()} is replaced by \code{\link[=gs4_example]{gs4_example()}}.

\code{sheets_examples()} is replaced by \code{\link[=gs4_examples]{gs4_examples()}}.

\code{sheets_browse()} is replaced by \code{\link[=gs4_browse]{gs4_browse()}}.

\code{sheets_find()} is replaced by \code{\link[=gs4_find]{gs4_find()}}.

\code{sheets_endpoints()} is replaced by \code{\link[=gs4_endpoints]{gs4_endpoints()}}.
}
\keyword{internal}
