% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{googlesheets4-deprecated}
\alias{googlesheets4-deprecated}
\alias{sheets_sheets}
\alias{sheets_cells}
\alias{sheets_read}
\alias{sheets_write}
\title{Deprecated functions}
\usage{
sheets_sheets(...)

sheets_cells(...)

sheets_read(...)

sheets_write(...)
}
\arguments{
\item{...}{Passed on to the function that succeeds the deprecated function.}
}
\description{
\lifecycle{deprecated}

These functions are deprecated and will be removed in a future release of
googlesheets4.

\code{sheets_sheets()} is replaced by \code{\link[=sheet_names]{sheet_names()}}.

\code{sheets_cells()} is replaced by \code{\link[=range_read_cells]{range_read_cells()}}.

\code{sheets_read()} is replaced by \code{\link[=range_read]{range_read()}} (which is a synonym for
\code{\link[=read_sheet]{read_sheet()}}).

\code{sheets_write()} is replaced by \code{\link[=sheet_write]{sheet_write()}} (which is a synonym for
\code{\link[=write_sheet]{write_sheet()}}).
}
\keyword{internal}
