% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{googlesheets4-deprecated}
\alias{googlesheets4-deprecated}
\alias{sheets_sheets}
\title{Deprecated functions}
\usage{
sheets_sheets(ss)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}
}
\description{
\lifecycle{deprecated}

These functions are deprecated and will be removed in a future release of
googlesheets4.

\code{sheets_sheets()} is replaced by \code{sheets_sheet_names()}.
}
\keyword{internal}
