% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets4-package.R
\docType{package}
\name{googlesheets4-package}
\alias{googlesheets4}
\alias{googlesheets4-package}
\title{googlesheets4: Access Google Sheets using the Sheets API v4}
\description{
Interact with Google Sheets through the Sheets API v4
    <https://developers.google.com/sheets/api>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverse/googlesheets4}
  \item Report bugs at \url{https://github.com/tidyverse/googlesheets4/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@rstudio.com} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
