% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auth.R
\name{sheets_auth_config}
\alias{sheets_auth_config}
\alias{sheets_api_key}
\alias{sheets_oauth_app}
\title{View or edit auth config}
\usage{
sheets_auth_config(app = NULL, path = NULL, api_key = NULL)

sheets_api_key()

sheets_oauth_app()
}
\arguments{
\item{app}{OAuth app. Defaults to a {AUTH_CONFIG_SOURCE} app.}

\item{path}{Path to the JSON file.}

\item{api_key}{API key. Defaults to a {AUTH_CONFIG_SOURCE} key. Necessary in
order to make unauthorized "token-free" requests for public resources.}
}
\value{
\code{sheets_auth_config()}: An object of R6 class \code{AuthState},
which is defined in the gargle package. \code{sheets_api_key()}: the
current API key. \code{sheets_oauth_app()}: the current
\code{\link[httr:oauth_app]{httr::oauth_app()}}.
}
\description{
These functions give the user more control over auth than what is
possible with \code{\link[=sheets_auth]{sheets_auth()}}. Learn more in Google's documentation:
\href{https://support.google.com/googleapi/answer/6158857?hl=en&ref_topic=7013279}{Credentials, access, security, and identity}
and \href{https://developers.google.com/identity/protocols/OAuth2InstalledApp}{Using OAuth 2.0 for Installed Applications}.
\code{sheets_auth_config()} gives control of:
\itemize{
\item The OAuth app. If you want to use your own app, setup a new project
in \href{https://console.developers.google.com}{Google Developers Console}.
Follow the instructions in \href{https://developers.google.com/identity/protocols/OAuth2InstalledApp}{OAuth 2.0 for Mobile & Desktop Apps}
to obtain your own client ID and secret. Either make an app from
your client ID and secret via \code{\link[httr:oauth_app]{httr::oauth_app()}} or provide a path
to the JSON file containing same, which you can download from
\href{https://console.developers.google.com}{Google Developers Console}.
\item The API key. If googlesheets4 is deauthorized via
\code{\link[=sheets_deauth]{sheets_deauth()}}, all requests will be sent with an API key in
lieu of a token. If you want to provide your own API key, setup a
project as described above and follow the instructions in
\href{https://support.google.com/googleapi/answer/6158862}{Setting up API keys}.
}

\code{sheets_api_key()} and \code{sheets_oauth_app()} retrieve the
currently configured API key and OAuth app, respectively.
}
\examples{
## retrieve current config
sheets_auth_config()

if (require(httr)) {
  ## bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  sheets_auth_config(app = google_app)
}

\dontrun{
## bring your own app via JSON downloaded from Google Developers Console
sheets_auth_config(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)

sheets_api_key()
sheets_oauth_app()

sheets_auth_config(api_key = "123")
sheets_api_key()
}
}
\seealso{
Other auth functions: \code{\link{sheets_auth}},
  \code{\link{sheets_deauth}}
}
\concept{auth functions}
