% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_browse.R
\name{sheets_browse}
\alias{sheets_browse}
\title{Visit Sheet in browser}
\usage{
sheets_browse(ss)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}
}
\value{
The Sheet's browser URL, invisibly.
}
\description{
Visits a Google Sheet in your default browser, if session is interactive.
}
\examples{
sheets_example("mini-gap") \%>\% sheets_browse()
}
