% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_cells.R
\name{sheets_cells}
\alias{sheets_cells}
\title{Read cells from a Sheet}
\usage{
sheets_cells(ss, sheet = NULL, range = NULL)
}
\arguments{
\item{ss}{Something that uniquely identifies a Google Sheet. Processed
through \code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
visible sheet. \emph{basically copied from readxl}}

\item{range}{A cell range to read from, as described in cell-specification
(does not link to anything yet) \emph{basically copied from readxl}}
}
\value{
something TBD
}
\description{
WIP WIP WIP
}
\examples{
x <- sheets_cells(sheets_example("design-dates"))
}
