% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_create.R
\name{sheets_create}
\alias{sheets_create}
\title{Create a new Sheet}
\usage{
sheets_create(name, ..., sheets = NULL)
}
\arguments{
\item{name}{The name of the spreadsheet.}

\item{...}{Optional spreadsheet properties that can be set through this API
endpoint, such as locale and time zone.}

\item{sheets}{Optional named list of data frames. One sheet is created for
each data frame.}
}
\value{
\itemize{
\item \code{sheets_get()}: A list with S3 class \code{sheets_Spreadsheet}, for printing
purposes.
\item \code{sheets_sheets()}: A character vector.
}
}
\description{
Creates an entirely new Sheet (spreadsheet or workbook). Offers some control
over the initial set of sheets (worksheets or tabs). CAUTION: this function
is still being developed and, for example, currently sends all data as
character.
}
\examples{
if (sheets_has_token()) {
  sheets_create("sheets-create-demo-1")

  sheets_create("sheets-create-demo-2", locale = "en_CA")

  sheets_create(
    "sheets-create-demo-3",
    locale = "fr_FR",
    timeZone = "Europe/Paris"
  )

  sheets_create(
    "sheets-create-demo-4",
    sheets = list(iris = head(iris), mtcars = head(mtcars))
  )

  # clean up
  sheets_find("sheets-create-demo") \%>\% googledrive::drive_trash()
}
}
\seealso{
Wraps the \code{spreadsheets.create} endpoint:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/create}
}
}
