% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_create.R
\name{sheets_create}
\alias{sheets_create}
\title{Create a new Sheet}
\usage{
sheets_create(name = sheets_random(), ..., sheets = NULL)
}
\arguments{
\item{name}{The name of the new spreadsheet.}

\item{...}{Optional spreadsheet properties that can be set through this API
endpoint, such as locale and time zone.}

\item{sheets}{Optional input for initializing (work)sheets. If unspecified,
the Sheets API automatically creates an empty "Sheet1". You can provide a
vector of sheet names, a data frame, or a (possibly named) list of data
frames. See the examples.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
\lifecycle{experimental}

Creates an entirely new (spread)Sheet (or, in Excel-speak, workbook).
Optionally, you can also provide names and/or data for the initial set of
(work)sheets. Any initial data provided via \code{sheets} is styled as a table,
as described in \code{\link[=sheets_write]{sheets_write()}}.
}
\examples{
if (sheets_has_token()) {
  sheets_create("sheets-create-demo-1")

  sheets_create("sheets-create-demo-2", locale = "en_CA")

  sheets_create(
    "sheets-create-demo-3",
    locale = "fr_FR",
    timeZone = "Europe/Paris"
  )

  sheets_create(
    "sheets-create-demo-4",
    sheets = c("alpha", "beta")
  )

  my_data <- data.frame(x = 1)
  sheets_create(
    "sheets-create-demo-5",
    sheets = my_data
  )

  sheets_create(
    "sheets-create-demo-6",
    sheets = list(iris = head(iris), mtcars = head(mtcars))
  )

  # clean up
  sheets_find("sheets-create-demo") \%>\% googledrive::drive_trash()
}
}
\seealso{
Wraps the \code{spreadsheets.create} endpoint:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/create}
}

There is an article on writing Sheets:
\itemize{
\item \url{https://googlesheets4.tidyverse.org/articles/articles/write-sheets.html}
}

Other write functions: 
\code{\link{range_flood}()},
\code{\link{sheets_append}()},
\code{\link{sheets_delete}()},
\code{\link{sheets_edit}()},
\code{\link{sheets_formula}()},
\code{\link{write_sheet}()}
}
\concept{write functions}
