% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_edit.R
\name{sheets_edit}
\alias{sheets_edit}
\title{(Over)write new data into a range}
\usage{
sheets_edit(ss, data, sheet = NULL, range = NULL, col_names = TRUE)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{data}{A data frame.}

\item{sheet}{Sheet to write into, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via \code{range}. If neither argument specifies the sheet, defaults to the first visible sheet.}

\item{range}{A cell range to read from. If \code{NULL}, all non-empty cells are
read. Otherwise specify \code{range} as described in \href{https://developers.google.com/sheets/api/guides/concepts#a1_notation}{Sheets A1 notation}
or using the helpers documented in \link{cell-specification}. Sheets uses
fairly standard spreadsheet range notation, although a bit different from
Excel. Examples of valid ranges: \code{"Sheet1!A1:B2"}, \code{"Sheet1!A:A"},
\code{"Sheet1!1:2"}, \code{"Sheet1!A5:A"}, \code{"A1:B2"}, \code{"Sheet1"}. Interpreted
strictly, even if the range forces the inclusion of leading, trailing, or
embedded empty rows or columns. Takes precedence over \code{skip}, \code{n_max} and
\code{sheet}. Note \code{range} can be a named range, like \code{"sales_data"}, without
any cell reference.}

\item{col_names}{Logical, indicating whether to send the column names of
\code{data}.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
\lifecycle{experimental}
Writes a data frame into a range. Main differences from \code{\link[=sheets_write]{sheets_write()}}:
\itemize{
\item The edited cells are not explicitly formatted or styled as a table.
Nothing special is done re: a header row or freezing rows.
\item (Work)sheet dimensions are not changed.
\item The target (spread)Sheet and (work)sheet must already exist. There is no
ability to create a Sheet or add a worksheet.
}
}
\examples{
if (sheets_has_token()) {
  df <- data.frame(
    x = 1:3,
    y = letters[1:3]
  )

  # create a Sheet with some initial, placeholder data
  (ss <- sheets_create("sheets-edit-demo", sheets = "alpha"))

  #  write df somewhere other than the "upper left corner"
  sheets_edit(ss, data = df, range = "D6")

  # view your magnificent creation in the browser
  # sheets_browse(ss)

  # clean up
  googledrive::drive_rm(ss)
}
}
