% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_example.R
\name{sheets_example}
\alias{sheets_example}
\alias{sheets_examples}
\title{File IDs of example Sheets}
\usage{
sheets_example(name = names(sheets_examples()))

sheets_examples()
}
\arguments{
\item{name}{Nickname of an example Sheet.}
}
\value{
\itemize{
\item \code{sheets_example()}: a single \link{sheets_id} object
\item \code{sheets_examples()}: a named character vector of all built-in examples
}
}
\description{
googlesheets4 ships with static IDs for some world-readable example Sheets
for use in examples and documentation. These functions make them easy to
access by a nickname.
}
\examples{
sheets_examples()

sheets_example("gapminder")
}
