% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_example.R
\name{sheets_example}
\alias{sheets_example}
\title{Access IDs of example Sheets}
\usage{
sheets_example(name = NULL)
}
\arguments{
\item{name}{Nickname of the example Sheet. If \code{NULL}, the examples are
listed.}
}
\value{
Either a single \link{sheets_id} object or a named character vector of all
built-in examples.
}
\description{
googlesheets4 ships with static IDs for some world-readable example Sheets
for use in examples and documentation. This function make them easy to
access.
}
\examples{
sheets_example()
sheets_example("gapminder")
}
