% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_get.R
\name{sheets_get}
\alias{sheets_get}
\title{Get metadata for a spreadsheet}
\usage{
sheets_get(ss)
}
\arguments{
\item{ss}{Something that uniquely identifies a Google Sheet. Processed
through \code{\link[=as_sheets_id]{as_sheets_id()}}.}
}
\value{
A list with S3 class \code{sheets_meta}, for printing purposes.
}
\description{
WIP. Name should maybe involve "meta" or "metadata"? OTOH there's an obvious
connection to \code{\link[googledrive:drive_get]{googledrive::drive_get()}}.
}
\examples{
sheets_get(sheets_example("design-dates"))
sheets_get(sheets_example("gapminder"))
sheets_get(sheets_example("mini-gap"))
sheets_get(sheets_example("ff"))
}
