% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_auth.R
\name{sheets_has_token}
\alias{sheets_has_token}
\title{Is there a token on hand?}
\usage{
sheets_has_token()
}
\value{
Logical.
}
\description{
Reports whether googlesheets4 has stored a token, ready for use in downstream
requests.
}
\examples{
sheets_has_token()
}
\seealso{
Other low-level API functions: \code{\link{request_generate}},
  \code{\link{request_make}}, \code{\link{sheets_token}}
}
\concept{low-level API functions}
