% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_list.R
\name{sheets_list}
\alias{sheets_list}
\title{List your Google Sheets}
\usage{
sheets_list(...)
}
\arguments{
\item{...}{Arguments (other than \code{type}, which is hard-wired as \code{type = "spreadsheet"}) that are passed along to \code{\link[googledrive:drive_find]{googledrive::drive_find()}}.}
}
\value{
An object of class \code{\link[googledrive]{dribble}}, a tibble with one row per item.
}
\description{
Lists your Google Sheets. This is a very thin wrapper around
\code{\link[googledrive:drive_find]{googledrive::drive_find()}}, that specifies you want to list Drive files
where \code{type = "spreadsheet"}.
}
\examples{
if (interactive()) {
  sheets_list()
}
}
