% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_sheet_copy.R
\name{sheets_sheet_copy}
\alias{sheets_sheet_copy}
\title{Copy one (work)sheet}
\usage{
sheets_sheet_copy(
  origin_ss,
  origin_sheet,
  destination_ss = origin_ss,
  destination_sheet = NULL
)
}
\arguments{
\item{origin_ss}{Something that identifies the origin Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{origin_sheet}{Sheet name or position of the sheet to copy.}

\item{destination_ss}{Optional, defaults to \code{origin_ss}.
Something that identifies the destination Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}, or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{destination_sheet}{Optional sheet name of the newly copied sheet..}
}
\value{
A list containing the name and id of the newly copied sheet.
}
\description{
Copy one (work)sheet from a (spread)Sheet to the same or a different (spread)Sheet
}
\examples{
if (sheets_has_token()) {
  ss <- sheets_create("copy-sheets-from-me")
  ss2 <- sheets_create("copy-sheets-to-me")
  sheets_sheet_add(ss, c("alpha", "beta", "gamma"))

  # get an overview of the sheets
  sheets_sheet_data(ss)

  # rename sheets
  sheets_sheet_copy(ss, "gamma", destination_sheet="gamma ray")
  sheets_sheet_copy(ss, "beta", ss2, "beta particle")

  # get an overview of the sheets
  sheets_sheet_data(ss)
  sheets_sheet_data(ss2)

  # cleanup
  sheets_find("copy-sheets-from-me") \%>\% googledrive::drive_rm()
  sheets_find("copy-sheets-to-me") \%>\% googledrive::drive_rm()
}
}
\seealso{
Makes an \code{copyTo} request:
\itemize{
\item \url{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.sheets/copyTo}
}

Other worksheet functions: 
\code{\link{sheets_sheet_add}()},
\code{\link{sheets_sheet_data}()},
\code{\link{sheets_sheet_delete}()},
\code{\link{sheets_sheet_rename}()}
}
\concept{worksheet functions}
