% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_write.R
\name{write_sheet}
\alias{write_sheet}
\alias{sheets_write}
\title{(Over)write new data into a worksheet}
\usage{
write_sheet(data, ss = NULL, sheet = NULL)

sheets_write(data, ss = NULL, sheet = NULL)
}
\arguments{
\item{data}{A data frame. If it has zero rows, we send one empty pseudo-row
of data, so that we can apply the usual table styling. This empty row goes
away (gets filled, actually) the first time you send more data with
\code{\link[=sheet_append]{sheet_append()}}.}

\item{ss}{Something that identifies a Google Sheet: its file ID, a URL from
which we can recover the ID, an instance of \code{googlesheets4_spreadsheet}
(returned by \code{\link[=sheets_get]{sheets_get()}}), or a \code{\link[googledrive:dribble]{dribble}}, which
is how googledrive represents Drive files. Processed through
\code{\link[=as_sheets_id]{as_sheets_id()}}.}

\item{sheet}{Sheet to write into, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number.}
}
\value{
The input \code{ss}, as an instance of \code{\link{sheets_id}}
}
\description{
\lifecycle{experimental}

Writes a data frame into a (work)sheet inside a (spread)Sheet. The target
sheet is styled as a table:
\itemize{
\item Special formatting is applied to the header row, which holds column
names.
\item The first row (header row) is frozen.
\item Sheet dimensions are set to "shrink wrap" the \code{data}.
}

If no existing Sheet is specified via \code{ss}, this function delegates to
\code{\link[=sheets_create]{sheets_create()}} and the new Sheet's name is randomly generated. If that's
undesirable, call \code{\link[=sheets_create]{sheets_create()}} directly to get more control.

If no \code{sheet} is specified or if \code{sheet} doesn't identify an existing sheet, a
new sheet is added to receive the \code{data}. If \code{sheet} specifies an existing
sheet, it is effectively overwritten! All pre-existing values, formats, and
dimensions are cleared and the targeted sheet gets new values and dimensions
from \code{data}.
}
\examples{
if (sheets_has_token()) {
  df <- data.frame(
    x = 1:3,
    y = letters[1:3]
  )

  # specify only a data frame, get a new Sheet, with a random name
  ss <- sheets_write(df)
  sheets_read(ss)

  # clean up
  googledrive::drive_rm(ss)

  # create a Sheet with some initial, placeholder data
  ss <- sheets_create(
    "sheets-write-demo",
    sheets = list(alpha = data.frame(x = 1), omega = data.frame(x = 1))
  )

  # write df into its own, new sheet
  sheets_write(df, ss = ss)

  # write mtcars into the sheet named "omega"
  sheets_write(mtcars, ss = ss, sheet = "omega")

  # get an overview of the sheets
  sheet_properties(ss)

  # view your magnificent creation in the browser
  # sheets_browse(ss)

  # clean up
  googledrive::drive_rm(ss)
}
}
\seealso{
Other write functions: 
\code{\link{range_delete}()},
\code{\link{range_flood}()},
\code{\link{range_write}()},
\code{\link{sheet_append}()},
\code{\link{sheets_create}()},
\code{\link{sheets_formula}()}
}
\concept{write functions}
